fixed some type issues

+++ src/main/java/RewardValue.java
@@ -12,8 +12,8 @@ public class RewardValue {
         this.milesValue = milesValue;
     }
 
-    public double convert_from_cash_to_miles() {
-        return cashValue / 0.0035;
+    public int convert_from_cash_to_miles() {
+        return (int) (cashValue / 0.0035);
     }
 
     public double convert_from_miles_to_cash() {
